+++ src/containers/studies/StudiesActions.js
@@ -5,10 +5,20 @@
 import { newRequestSequence } from 'redux-reqseq';
 import type { RequestSequence } from 'redux-reqseq';
 
+const ADD_PARTICIPANT :'ADD_PARTICIPANT' = 'ADD_PARTICIPANT';
+const addStudyParticipant :RequestSequence = newRequestSequence(ADD_PARTICIPANT);
+
 const GET_STUDIES :'GET_STUDIES' = 'GET_STUDIES';
 const getStudies :RequestSequence = newRequestSequence(GET_STUDIES);
 
+const CREATE_STUDY :'CREATE_STUDY' = 'CREATE_STUDY';
+const createStudy :RequestSequence = newRequestSequence(CREATE_STUDY);
+
 export {
+  ADD_PARTICIPANT,
+  CREATE_STUDY,
   GET_STUDIES,
+  addStudyParticipant,
+  createStudy,
   getStudies,
 };
