+++ src/containers/studies/StudiesContainer.js
@@ -20,11 +20,10 @@ import type { RequestSequence, RequestState } from 'redux-reqseq';
 
 import CreateStudyModal from './components/CreateStudyModal';
 import StudyCard from './components/StudyCard';
-import { GET_STUDIES, getStudies } from './StudiesActions';
+import { CREATE_STUDY, GET_STUDIES, getStudies } from './StudiesActions';
 
 import * as RoutingActions from '../../core/router/RoutingActions';
 import { resetRequestState } from '../../core/redux/ReduxActions';
-import { CREATE_STUDY } from '../study/StudyActions';
 
 const { OPENLATTICE_ID_FQN } = Constants;
 
@@ -148,7 +147,7 @@ class StudiesContainer extends Component<Props, State> {
 const mapStateToProps = (state :Map) => ({
   requestStates: {
     [GET_STUDIES]: state.getIn(['studies', GET_STUDIES, 'requestState']),
-    [CREATE_STUDY]: state.getIn(['study', CREATE_STUDY, 'requestState']),
+    [CREATE_STUDY]: state.getIn(['studies', CREATE_STUDY, 'requestState']),
 
   },
   studies: state.getIn(['studies', 'studies'])
